+++ packages/core/src/objects/object-stream-extractor/dict/object-stream-dict.validation.test.ts
@@ -1,5 +1,5 @@
 import { assert, expect, test } from "vitest";
-import { makeDict } from "../body/object-stream-body.test-helpers";
+import { makeDict } from "../body/object-stream-body.test.helpers";
 import { ObjectStreamDict } from "./index";
 
 test("validateは正しい辞書から/First,/N,needsDecompressを取得できる", () => {
